securityDeposit from $$ to month

+++ modules/koki-dto/src/main/kotlin/com/wutsi/koki/listing/dto/UpdateListingLeasingRequest.kt
@@ -4,5 +4,5 @@ data class UpdateListingLeasingRequest(
     val leaseTerm: Int? = null,
     val noticePeriod: Int? = null,
     val advanceRent: Int? = null,
-    val securityDeposit: Long? = null,
+    val securityDeposit: Int? = null,
 )
